Now merge layers of contiguous files when opening them, or open files as separate layers if they are on a different transect. Possibility to open twice the same file to compare results of different algo etc... Some improvement needed as for speeding the opening file process...
Former-commit-id: 58ef71bdf1c2520d75cb22cfd4582526be30ef33

+++ @ac_data_cl/get_datamat.m
@@ -8,7 +8,7 @@ function datamat=get_datamat(data,field)
 % end
 varlist=who(data.MatfileData);
 if nansum(strcmpi(varlist,(deblank(field))))==1
-    datamat=data.MatfileData.(lower(deblank(field)));
+    datamat=double(data.MatfileData.(lower(deblank(field))));
 else
     datamat=[];
 end
